refactor: Address reviewer feedback on TensorDict fix
- Move TensorDict tests from standalone file to test_data_collator.py
- Simplify comments from verbose explanation to short reference: 'Call .keys() explicitly to avoid issue #42370'
- Delete tests/trainer/test_tensordict_compatibility.py (tests now in test_data_collator.py)

Addresses feedback from @ligz08 in PR review
