+++ src/transformers/feature_extraction_sequence_utils.py
@@ -122,8 +122,7 @@ class SequenceFeatureExtractor(FeatureExtractionMixin):
         # If we have a list of dicts, let's convert it in a dict of lists
         # We do this to allow using this method as a collate_fn function in PyTorch Dataloader
         if isinstance(processed_features, (list, tuple)) and isinstance(processed_features[0], (dict, BatchFeature)):
-            # Use .keys() explicitly to support dict-like objects (e.g., TensorDict) that implement
-            # __iter__ differently than standard dicts (e.g., iterating over batch dimensions)
+            # Call .keys() explicitly to avoid issue #42370
             processed_features = {
                 key: [example[key] for example in processed_features] for key in processed_features[0].keys()
             }
