+++ src/transformers/models/mluke/tokenization_mluke.py
@@ -1287,8 +1287,7 @@ class MLukeTokenizer(PreTrainedTokenizer):
         # If we have a list of dicts, let's convert it in a dict of lists
         # We do this to allow using this method as a collate_fn function in PyTorch Dataloader
         if isinstance(encoded_inputs, (list, tuple)) and isinstance(encoded_inputs[0], Mapping):
-            # Use .keys() explicitly to support dict-like objects (e.g., TensorDict) that implement
-            # __iter__ differently than standard dicts (e.g., iterating over batch dimensions)
+            # Call .keys() explicitly to avoid issue #42370
             encoded_inputs = {key: [example[key] for example in encoded_inputs] for key in encoded_inputs[0].keys()}
 
         # The model's main input name, usually `input_ids`, has be passed for padding
