+++ src/transformers/tokenization_utils_base.py
@@ -3481,8 +3481,7 @@ class PreTrainedTokenizerBase(SpecialTokensMixin, PushToHubMixin):
             and len(encoded_inputs) > 0
             and isinstance(encoded_inputs[0], Mapping)
         ):
-            # Use .keys() explicitly to support dict-like objects (e.g., TensorDict) that implement
-            # __iter__ differently than standard dicts (e.g., iterating over batch dimensions)
+            # Call .keys() explicitly to avoid issue #42370
             encoded_inputs = {key: [example[key] for example in encoded_inputs] for key in encoded_inputs[0].keys()}
 
         # The model's main input name, usually `input_ids`, has been passed for padding
